requirements.txt
openai>=1.40.0
chromadb>=0.5.5
streamlit>=1.36.0
requests>=2.32.0
beautifulsoup4>=4.12.0
trafilatura>=1.12.0
python-dotenv>=1.0.1
tqdm>=4.66.0
pytest>=8.2.0
